fix: "error deleting object from bucket "test-db-manager-bucket" using key "": operation error S3: DeleteObject, serialization failed: serialization failed: input member Key must not be empty"

+++ pkg/database/service.go
@@ -178,9 +178,11 @@ func (s service) Delete(id uint) error {
 	}
 
 	key := strings.TrimPrefix(u.Path, "/")
-	err = s.s3Client.Delete(s.c.Bucket, key)
-	if err != nil {
-		return err
+	if key != "" {
+		err = s.s3Client.Delete(s.c.Bucket, key)
+		if err != nil {
+			return err
+		}
 	}
 
 	return s.repository.Delete(id)
